Add improvements for DI
